Add markup share
The criticism of the prior markup forwarding mechanism was that someone
could take over the ad at a very low markup with their first address and
then increase the markup with their second address, basically sending the
markup increase from the second address to their first, leaving the
original seller with the low markup increase as a gain.

An example:

- Alice buys the ad for 1 ETH and lets it run for half a year. It is now
  worth 0.5 ETH
- Bob comes along with his first address A1 and buys the ad from Alice
  for 0.500000...1 ETH, Alice gets sent 1 Wei
- Bob then buys the ad immediately again but with his second address A2
  for 1 ETH at a markup of 0.499999...9 ETH, sending this markup to his
  first address A1.
- Alice feels cheated by Bob's behavior as she expects to receive
  0.49999...9 ETH, the actual markup Bob contributed to her property.

So how can we fairly forward the price increase of the property to the
prior controllers without the above described issue?

Authoritatively, we know that controllers of the ad have the incentive
to minimize their tax burden. At the same time, the legit use case is
that the ad's primary utility is to advertise on Kiwi News. Hence, we
can assume that the total taxes paid per controller period are a great
indicator for who paid who much markup for how long. Here are the
changes we propose:

1. All the markups are added to a shared pool.
2. Upon selling the ad to a new controller, the seller's tax burden is
   recorded.
3. All former controllers can call a withdrawal function that allows
   them to take funds out of the shared pool of markups.